prints the size of various variables

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ long long int x;
 char y;
 float z;
 
-printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(y));
-printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(v));
-printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(w));
-printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(y));
-printf("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(z));
+printf("Size of a char: %zu byte(s)\n", sizeof(y));
+printf("Size of an int: %zu byte(s)\n", sizeof(v));
+printf("Size of a long int: %zu byte(s)\n", sizeof(w));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(y));
+printf("Size of a float: %zu byte(s)\n", sizeof(z));
 return (0);
 
 }
